Don't crash when trying to display large arrays
Even though the custom graphic object is displayed, the text editor still called the getText() method of ValueRenderer. When the array was very large, this caused .formatted() to be called on the array, reading to an ut of memory. This happened even though the text is never used. The solution is to not render large arrays and instead display the same summary description as is shown in the graphical renderer.