Monadic compose cannot be inversed
This is due to the change of behaviour of monadic compose. Previously the left function was called monadically with the result from the right function. After the change the left function is called dyadically with the original argument to the left (the result of the right argument is passed on the right). This means that the argument is used twice which prevents derivation of an inverse function.