Raise a parse error on invalid operator usage
Previously, operators were only identified when explicitly
given to the right of a function. In other circumstances
they were simply parsed as a variable, resulting is unexpected
behaviour when making a typo. In such cases, the user would
get a "variable undefined" error, with the variable name
being given as the name of the operator.